extract some code to functions in discovery helper
