+++ src/helpers/index.js
@@ -5,5 +5,5 @@ export * from './getPrice'
 export { default as getPriceRangeFromStocks } from './getPriceRangeFromStocks'
 export * from './getQueryParams'
 export * from './getShareURL'
-export { default as isRecommendationFinished } from './isRecommendationFinished'
+export * from './discovery'
 export { default as isSameDayInEachTimezone } from './isSameDayInEachTimezone'
